generalizing database classes and calls
